+++ src/conic/cop.rs
@@ -26,8 +26,9 @@ impl Default for Cop {
 
 impl Cop {
 
+  // default theta1 = theta2 = 45 deg
   pub fn new() -> Self {
-    Self::from_params(0.0, 0.0)
+    Self::from_params(HALF_PI.half(), 0.0)
   }
 
   pub fn from_params(theta_a: f64, nu: f64) -> Self {
@@ -71,7 +72,7 @@ impl CanonicalProjection for Cop {
   
   fn proj(&self, xyz: &XYZ) -> Option<ProjXY> {
     if self.is_lat_in_domain_of_validity(xyz.z) {
-      let lon = xyz.y.atan2(xyz.z);
+      let lon = xyz.y.atan2(xyz.x);
       // r(dec) depends on tan(dec - theta_a) 
       // => -pi/2 < dec - theta_a < pi/2 => -pi/2 + theta_a < dec < pi/2 + theta
       let r = xyz.z / (1.0 - xyz.z.pow2()).sqrt();
@@ -95,7 +96,7 @@ impl CanonicalProjection for Cop {
     let y2d = self.y0 - pos.y;
     let r2 = x2d.pow2() + y2d.pow2();
     let r = if self.conic.negative_ta { -(r2.sqrt()) } else { r2.sqrt() };
-    let lon =  y2d.atan2(x2d) / self.c; // no need to divide both y2d and x2d by r
+    let lon =  (x2d / r).atan2(y2d / r) / self.c; // / r important because of its sign
     if (-PI - EPS..PI + EPS).contains(&lon) {
       // dec = this.ta + atan(this.cotanta - r / this.cosnu);
       // sin(a+b) = sin(a)cos(b) + cos(a)sin(b)